requirements.txt
Pillow
numpy
torch